Move some scaffolder code to shared location for re-use with other database engine
