+++ src/Core/RevEng.Core.60/Routines/FunctionScaffolder.cs
@@ -1,0 +1,88 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using Microsoft.EntityFrameworkCore.Design;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Scaffolding;
+using RevEng.Core.Abstractions;
+using RevEng.Core.Abstractions.Metadata;
+
+namespace RevEng.Core.Modules
+{
+    public abstract class FunctionScaffolder : IRoutineScaffolder
+    {
+#pragma warning disable SA1401 // Fields should be private
+        internal readonly ICSharpHelper Code;
+        internal IndentedStringBuilder Sb;
+#pragma warning restore SA1401 // Fields should be private
+
+        protected FunctionScaffolder([System.Diagnostics.CodeAnalysis.NotNull] ICSharpHelper code)
+        {
+            ArgumentNullException.ThrowIfNull(code);
+
+            this.Code = code;
+        }
+
+        public string FileNameSuffix { get; set; }
+
+        public SavedModelFiles Save(ScaffoldedModel scaffoldedModel, string outputDir, string nameSpaceValue, bool useAsyncCalls)
+        {
+            ArgumentNullException.ThrowIfNull(scaffoldedModel);
+
+            Directory.CreateDirectory(outputDir);
+
+            var contextPath = Path.GetFullPath(Path.Combine(outputDir, scaffoldedModel.ContextFile.Path));
+            var path = Path.GetDirectoryName(contextPath);
+            if (path != null)
+            {
+                Directory.CreateDirectory(path);
+                File.WriteAllText(contextPath, scaffoldedModel.ContextFile.Code, Encoding.UTF8);
+            }
+
+            var additionalFiles = new List<string>();
+
+            foreach (var entityTypeFile in scaffoldedModel.AdditionalFiles)
+            {
+                var additionalFilePath = Path.Combine(outputDir, entityTypeFile.Path);
+                var addpath = Path.GetDirectoryName(additionalFilePath);
+                if (addpath != null)
+                {
+                    Directory.CreateDirectory(addpath);
+                    File.WriteAllText(additionalFilePath, entityTypeFile.Code, Encoding.UTF8);
+                    additionalFiles.Add(additionalFilePath);
+                }
+            }
+
+            return new SavedModelFiles(contextPath, additionalFiles);
+        }
+
+        public ScaffoldedModel ScaffoldModel(RoutineModel model, ModuleScaffolderOptions scaffolderOptions, List<string> schemas, ref List<string> errors)
+        {
+            ArgumentNullException.ThrowIfNull(model);
+
+            ArgumentNullException.ThrowIfNull(errors);
+
+            ArgumentNullException.ThrowIfNull(scaffolderOptions);
+
+            var result = new ScaffoldedModel();
+
+            errors.AddRange(model.Errors);
+
+            schemas = schemas ?? new List<string>();
+
+            var dbContext = WriteDbContext(scaffolderOptions, model, schemas.Distinct().ToList());
+
+            result.ContextFile = new ScaffoldedFile
+            {
+                Code = dbContext,
+                Path = Path.GetFullPath(Path.Combine(scaffolderOptions.ContextDir, scaffolderOptions.ContextName + $"{FileNameSuffix}.cs")),
+            };
+
+            return result;
+        }
+
+        protected abstract string WriteDbContext(ModuleScaffolderOptions scaffolderOptions, RoutineModel model, List<string> schemas);
+    }
+}
